docs: deploying to docs.vmware.com (#49)
The TOC works as is there and recognizes markdown sub-lists.
Changed the object extensibility schema links in the document to point to a zip holding the files. Going forward, we should follow this pattern to make the publishing and the experience reading the documentation streamlined and consistent. The publishing will be configured to include zip archives and links will cause the browser to download them.

Signed-off-by: Ivan Iliev <iviliev@vmware.com>